refactor(carRaceService): carRaceService 생성
