+++ src/main/java/carracing/dto/CarRaceRequest.java
@@ -1,0 +1,25 @@
+package carracing.dto;
+
+import carracing.Cars;
+import carracing.Track;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class CarRaceRequest {
+    private final List<String> carNames = new ArrayList<>();
+    private final int movingCount;
+
+    public CarRaceRequest(List<String> carNames, int movingCount) {
+        this.carNames.addAll(carNames);
+        this.movingCount = movingCount;
+    }
+
+    public Track toTrack() {
+        return new Track(Cars.makeCars(carNames));
+    }
+
+    public Integer getMovingCount() {
+        return movingCount;
+    }
+}
